clean up and deduplicate webdav MKCOL code
