+++ internal/http/services/owncloud/ocdav/mkcol.go
@@ -19,17 +19,18 @@
 package ocdav
 
 import (
-	"io"
+	"context"
 	"net/http"
 	"path"
 
 	rpc "github.com/cs3org/go-cs3apis/cs3/rpc/v1beta1"
 	provider "github.com/cs3org/go-cs3apis/cs3/storage/provider/v1beta1"
 	"github.com/cs3org/reva/pkg/appctx"
+	"github.com/rs/zerolog"
 	"go.opencensus.io/trace"
 )
 
-func (s *svc) handleMkcol(w http.ResponseWriter, r *http.Request, ns string) {
+func (s *svc) handlePathMkcol(w http.ResponseWriter, r *http.Request, ns string) {
 	ctx := r.Context()
 	ctx, span := trace.StartSpan(ctx, "mkcol")
 	defer span.End()
@@ -38,29 +39,54 @@ func (s *svc) handleMkcol(w http.ResponseWriter, r *http.Request, ns string) {
 
 	sublog := appctx.GetLogger(ctx).With().Str("path", fn).Logger()
 
-	buf := make([]byte, 1)
-	_, err := r.Body.Read(buf)
-	if err != io.EOF {
-		sublog.Error().Err(err).Msg("error reading request body")
-		w.WriteHeader(http.StatusUnsupportedMediaType)
-		return
+	ref := &provider.Reference{
+		Spec: &provider.Reference_Path{Path: fn},
 	}
 
-	client, err := s.getClient()
+	s.handleMkcol(ctx, w, r, ref, sublog)
+}
+
+func (s *svc) handleSpacesMkCol(w http.ResponseWriter, r *http.Request, spaceID string) {
+	ctx := r.Context()
+	ctx, span := trace.StartSpan(ctx, "spaces_mkcol")
+	defer span.End()
+
+	sublog := appctx.GetLogger(ctx).With().Str("path", r.URL.Path).Str("spaceid", spaceID).Str("handler", "mkcol").Logger()
+
+	ref, rpcStatus, err := s.lookUpStorageSpaceReference(ctx, spaceID, r.URL.Path)
 	if err != nil {
-		sublog.Error().Err(err).Msg("error getting grpc client")
+		sublog.Error().Err(err).Msg("error sending a grpc request")
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
 
-	// check fn exists
-	ref := &provider.Reference{
-		Spec: &provider.Reference_Path{Path: fn},
+	if rpcStatus.Code != rpc.Code_CODE_OK {
+		HandleErrorStatus(&sublog, w, rpcStatus)
+		return
 	}
-	statReq := &provider.StatRequest{Ref: ref}
-	statRes, err := client.Stat(ctx, statReq)
+
+	s.handleMkcol(ctx, w, r, ref, sublog)
+
+}
+
+func (s *svc) handleMkcol(ctx context.Context, w http.ResponseWriter, r *http.Request, ref *provider.Reference, log zerolog.Logger) {
+	if r.Body != http.NoBody {
+		w.WriteHeader(http.StatusUnsupportedMediaType)
+		return
+	}
+
+	gatewayClient, err := s.getClient()
 	if err != nil {
-		sublog.Error().Err(err).Msg("error sending a grpc stat request")
+		log.Error().Err(err).Msg("error getting grpc client")
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+
+	// check if ref exists
+	statReq := &provider.StatRequest{Ref: ref}
+	statRes, err := gatewayClient.Stat(ctx, statReq)
+	if err != nil {
+		log.Error().Err(err).Msg("error sending a grpc stat request")
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
@@ -69,15 +95,15 @@ func (s *svc) handleMkcol(w http.ResponseWriter, r *http.Request, ns string) {
 		if statRes.Status.Code == rpc.Code_CODE_OK {
 			w.WriteHeader(http.StatusMethodNotAllowed) // 405 if it already exists
 		} else {
-			HandleErrorStatus(&sublog, w, statRes.Status)
+			HandleErrorStatus(&log, w, statRes.Status)
 		}
 		return
 	}
 
 	req := &provider.CreateContainerRequest{Ref: ref}
-	res, err := client.CreateContainer(ctx, req)
+	res, err := gatewayClient.CreateContainer(ctx, req)
 	if err != nil {
-		sublog.Error().Err(err).Msg("error sending create container grpc request")
+		log.Error().Err(err).Msg("error sending create container grpc request")
 		w.WriteHeader(http.StatusInternalServerError)
 		return
 	}
@@ -85,9 +111,9 @@ func (s *svc) handleMkcol(w http.ResponseWriter, r *http.Request, ns string) {
 	case rpc.Code_CODE_OK:
 		w.WriteHeader(http.StatusCreated)
 	case rpc.Code_CODE_NOT_FOUND:
-		sublog.Debug().Str("path", fn).Interface("status", statRes.Status).Msg("conflict")
+		log.Debug().Str("path", r.URL.Path).Interface("status", statRes.Status).Msg("conflict")
 		w.WriteHeader(http.StatusConflict)
 	default:
-		HandleErrorStatus(&sublog, w, res.Status)
+		HandleErrorStatus(&log, w, res.Status)
 	}
 }
