+++ internal/http/services/owncloud/ocdav/spaces.go
@@ -156,72 +156,6 @@ func (s *svc) lookUpStorageSpaceReference(ctx context.Context, spaceID string, r
 	}, lSSRes.Status, nil
 }
 
-func (s *svc) handleSpacesMkCol(w http.ResponseWriter, r *http.Request, spaceID string) {
-	ctx := r.Context()
-	ctx, span := trace.StartSpan(ctx, "mkcol")
-	defer span.End()
-
-	sublog := appctx.GetLogger(ctx).With().Str("path", r.URL.Path).Str("spaceid", spaceID).Str("handler", "mkcol").Logger()
-
-	if r.Body != http.NoBody {
-		w.WriteHeader(http.StatusUnsupportedMediaType)
-		return
-	}
-
-	ref, rpcStatus, err := s.lookUpStorageSpaceReference(ctx, spaceID, r.URL.Path)
-	if err != nil {
-		sublog.Error().Err(err).Msg("error sending a grpc request")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	if rpcStatus.Code != rpc.Code_CODE_OK {
-		HandleErrorStatus(&sublog, w, rpcStatus)
-		return
-	}
-
-	gatewayClient, err := s.getClient()
-	if err != nil {
-		sublog.Error().Err(err).Msg("error getting grpc client")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	statReq := &storageProvider.StatRequest{Ref: ref}
-	statRes, err := gatewayClient.Stat(ctx, statReq)
-	if err != nil {
-		sublog.Error().Err(err).Msg("error sending a grpc stat request")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-
-	if statRes.Status.Code != rpc.Code_CODE_NOT_FOUND {
-		if statRes.Status.Code == rpc.Code_CODE_OK {
-			w.WriteHeader(http.StatusMethodNotAllowed) // 405 if it already exists
-		} else {
-			HandleErrorStatus(&sublog, w, statRes.Status)
-		}
-		return
-	}
-
-	req := &storageProvider.CreateContainerRequest{Ref: ref}
-	res, err := gatewayClient.CreateContainer(ctx, req)
-	if err != nil {
-		sublog.Error().Err(err).Msg("error sending create container grpc request")
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	switch res.Status.Code {
-	case rpc.Code_CODE_OK:
-		w.WriteHeader(http.StatusCreated)
-	case rpc.Code_CODE_NOT_FOUND:
-		sublog.Debug().Str("path", r.URL.Path).Interface("status", statRes.Status).Msg("conflict")
-		w.WriteHeader(http.StatusConflict)
-	default:
-		HandleErrorStatus(&sublog, w, res.Status)
-	}
-}
-
 func (s *svc) handleSpacesMove(w http.ResponseWriter, r *http.Request, srcSpaceID string) {
 	ctx := r.Context()
 	ctx, span := trace.StartSpan(ctx, "move")
